# internal/model/LQS.go
package model

type L21 struct {
	ID           string `json:"id" eorm:"id"`
	R路线编号        string `json:"路线编号" eorm:"路线编号"`
	R所在行政区划代码    string `json:"所在行政区划代码" eorm:"所在行政区划代码"`
	R路线名称        string `json:"路线名称" eorm:"路线名称"`
	R起点名称        string `json:"起点名称" eorm:"起点名称"`
	R止点名称        string `json:"止点名称" eorm:"止点名称"`
	R起点桩号        string `json:"起点桩号" eorm:"起点桩号"`
	R止点桩号        string `json:"止点桩号" eorm:"止点桩号"`
	R里程公里        string `json:"里程(公里)" eorm:"里程(公里)"`
	R技术等级代码      string `json:"技术等级代码" eorm:"技术等级代码"`
	R技术等级        string `json:"技术等级" eorm:"技术等级"`
	R是否一幅高速      string `json:"是否一幅高速" eorm:"是否一幅高速"`
	R车道数量个       string `json:"车道数量(个)" eorm:"车道数量(个)"`
	R面层类型代码      string `json:"面层类型代码" eorm:"面层类型代码"`
	R面层类型        string `json:"面层类型" eorm:"面层类型"`
	R路基宽度米       string `json:"路基宽度(米)" eorm:"路基宽度(米)"`
	R路面宽度米       string `json:"路面宽度(米)" eorm:"路面宽度(米)"`
	R面层厚度厘米      string `json:"面层厚度(厘米)" eorm:"面层厚度(厘米)"`
	R设计时速公里小时    string `json:"设计时速(公里/小时)" eorm:"设计时速(公里/小时)"`
	R修建年度        string `json:"修建年度" eorm:"修建年度"`
	R改建年度        string `json:"改建年度" eorm:"改建年度"`
	R最近一次修复养护年度  string `json:"最近一次修复养护年度" eorm:"最近一次修复养护年度"`
	R断链类型        string `json:"断链类型" eorm:"断链类型"`
	R是否城管路段      string `json:"是否城管路段" eorm:"是否城管路段"`
	R是否断头路段      string `json:"是否断头路段" eorm:"是否断头路段"`
	R路段收费性质      string `json:"路段收费性质" eorm:"路段收费性质"`
	R重复路段线路编号    string `json:"重复路段线路编号" eorm:"重复路段线路编号"`
	R重复路段起点桩号    string `json:"重复路段起点桩号" eorm:"重复路段起点桩号"`
	R重复路段终点桩号    string `json:"重复路段终点桩号" eorm:"重复路段终点桩号"`
	R养护里程        string `json:"养护里程" eorm:"养护里程"`
	R可绿化里程       string `json:"可绿化里程" eorm:"可绿化里程"`
	R已绿化里程       string `json:"已绿化里程" eorm:"已绿化里程"`
	R地貌代码        string `json:"地貌代码" eorm:"地貌代码"`
	R地貌汉字        string `json:"地貌汉字" eorm:"地貌汉字"`
	R涵洞数量个       string `json:"涵洞数量(个)" eorm:"涵洞数量(个)"`
	R管养单位名称      string `json:"管养单位名称" eorm:"管养单位名称"`
	R省际出入口       string `json:"省际出入口" eorm:"省际出入口"`
	R国道调整前路线编号   string `json:"国道调整前路线编号" eorm:"国道调整前路线编号"`
	R是否按干线公路管理接养 string `json:"是否按干线公路管理接养" eorm:"是否按干线公路管理接养"`
	R备注          string `json:"备注" eorm:"备注"`
}

type L25 struct {
	ID          string `json:"id" eorm:"id"`
	S隧道名称       string `json:"隧道名称" eorm:"隧道名称"`
	S隧道代码       string `json:"隧道代码" eorm:"隧道代码"`
	S隧道中心桩号     string `json:"隧道中心桩号" eorm:"隧道中心桩号"`
	S所属路线编号     string `json:"所属路线编号" eorm:"所属路线编号"`
	S所属路线名称     string `json:"所属路线名称" eorm:"所属路线名称"`
	S所属线路技术等级   string `json:"所属线路技术等级" eorm:"所属线路技术等级"`
	S隧道长度米      string `json:"隧道长度(米)" eorm:"隧道长度(米)"`
	S隧道净宽米      string `json:"隧道净宽(米)" eorm:"隧道净宽(米)"`
	S隧道净高米      string `json:"隧道净高(米)" eorm:"隧道净高(米)"`
	S隧道按长度分类代码  string `json:"隧道按长度分类代码" eorm:"隧道按长度分类代码"`
	S隧道按长度分类    string `json:"隧道按长度分类" eorm:"隧道按长度分类"`
	S是否水下隧道     string `json:"是否水下隧道" eorm:"是否水下隧道"`
	S修建年度       string `json:"修建年度" eorm:"修建年度"`
	S建设单位名称     string `json:"建设单位名称" eorm:"建设单位名称"`
	S设计单位名称     string `json:"设计单位名称" eorm:"设计单位名称"`
	S施工单位名称     string `json:"施工单位名称" eorm:"施工单位名称"`
	S监理单位名称     string `json:"监理单位名称" eorm:"监理单位名称"`
	S建成通车时间     string `json:"建成通车时间" eorm:"建成通车时间"`
	S隧道养护等级     string `json:"隧道养护等级" eorm:"隧道养护等级"`
	S管养单位性质代码   string `json:"管养单位性质代码" eorm:"管养单位性质代码"`
	S管养名称       string `json:"管养名称" eorm:"管养名称"`
	S监管单位名称     string `json:"监管单位名称" eorm:"监管单位名称"`
	S技术状况评定等级   string `json:"技术状况评定等级" eorm:"技术状况评定等级"`
	S技术状况评定日期   string `json:"技术状况评定日期" eorm:"技术状况评定日期"`
	S技术状况评定单位   string `json:"技术状况评定单位" eorm:"技术状况评定单位"`
	S土建结构评定等级   string `json:"土建结构评定等级" eorm:"土建结构评定等级"`
	S土建结构评定日期   string `json:"土建结构评定日期" eorm:"土建结构评定日期"`
	S土建结构评定单位   string `json:"土建结构评定单位" eorm:"土建结构评定单位"`
	S机电设施评定等级   string `json:"机电设施评定等级" eorm:"机电设施评定等级"`
	S机电设施评定日期   string `json:"机电设施评定日期" eorm:"机电设施评定日期"`
	S机电设施评定单位   string `json:"机电设施评定单位" eorm:"机电设施评定单位"`
	S其他工程设施评定等级 string `json:"其他工程设施评定等级" eorm:"其他工程设施评定等级"`
	S其他工程设施评定日期 string `json:"其他工程设施评定日期" eorm:"其他工程设施评定日期"`
	S其他工程设施评定单位 string `json:"其他工程设施评定单位" eorm:"其他工程设施评定单位"`
	S最近一次改造年度   string `json:"最近一次改造年度" eorm:"最近一次改造年度"`
	S最近一次改造完工日期 string `json:"最近一次改造完工日期" eorm:"最近一次改造完工日期"`
	S最近一次改造部位   string `json:"最近一次改造部位" eorm:"最近一次改造部位"`
	S最近一次改造工程性质 string `json:"最近一次改造工程性质" eorm:"最近一次改造工程性质"`
	S当前主要病害部位   string `json:"当前主要病害部位" eorm:"当前主要病害部位"`
	S当前主要病害描述   string `json:"当前主要病害描述" eorm:"当前主要病害描述"`
	S政区代码       string `json:"政区代码" eorm:"政区代码"`
	S是否在长大隧道目录中 string `json:"是否在长大隧道目录中" eorm:"是否在长大隧道目录中"`
	S备注         string `json:"备注" eorm:"备注"`
}

type L24 struct {
	ID             string `json:"id" eorm:"id"`
	Q桥梁名称          string `json:"桥梁名称" eorm:"桥梁名称"`
	Q桥梁代码          string `json:"桥梁代码" eorm:"桥梁代码"`
	Q桥梁中心桩号        string `json:"桥梁中心桩号" eorm:"桥梁中心桩号"`
	Q路线编号          string `json:"路线编号" eorm:"路线编号"`
	Q路线名称          string `json:"路线名称" eorm:"路线名称"`
	Q技术等级          string `json:"技术等级" eorm:"技术等级"`
	Q桥梁全长米         string `json:"桥梁全长(米)" eorm:"桥梁全长(米)"`
	Q跨径总长米         string `json:"跨径总长(米)" eorm:"跨径总长(米)"`
	Q单孔最大跨径米       string `json:"单孔最大跨径(米)" eorm:"单孔最大跨径(米)"`
	Q跨径组合孔米        string `json:"跨径组合(孔*米)" eorm:"跨径组合(孔*米)"`
	Q桥梁全宽米         string `json:"桥梁全宽(米)" eorm:"桥梁全宽(米)"`
	Q桥面净宽米         string `json:"桥面净宽(米)" eorm:"桥面净宽(米)"`
	Q按跨径分类代码       string `json:"按跨径分类代码" eorm:"按跨径分类代码"`
	Q按跨径分类类型       string `json:"按跨径分类类型" eorm:"按跨径分类类型"`
	Q按使用年限分类代码     string `json:"按使用年限分类代码" eorm:"按使用年限分类代码"`
	Q按使用年限分类类型     string `json:"按使用年限分类类型" eorm:"按使用年限分类类型"`
	Q主桥上部构造结构类型代码  string `json:"主桥上部构造结构类型代码" eorm:"主桥上部构造结构类型代码"`
	Q主桥上部构造结构类型类型  string `json:"主桥上部构造结构类型类型" eorm:"主桥上部构造结构类型类型"`
	Q主桥上部构造材料代码    string `json:"主桥上部构造材料代码" eorm:"主桥上部构造材料代码"`
	Q主桥上部构造材料名称    string `json:"主桥上部构造材料名称" eorm:"主桥上部构造材料名称"`
	Q桥墩类型代码        string `json:"桥墩类型代码" eorm:"桥墩类型代码"`
	Q桥墩类型类型        string `json:"桥墩类型类型" eorm:"桥墩类型类型"`
	Q设计荷载等级代码      string `json:"设计荷载等级代码" eorm:"设计荷载等级代码"`
	Q设计荷载等级        string `json:"设计荷载等级" eorm:"设计荷载等级"`
	Q抗震等级代码        string `json:"抗震等级代码" eorm:"抗震等级代码"`
	Q抗震等级          string `json:"抗震等级" eorm:"抗震等级"`
	Q跨越地物代码        string `json:"跨越地物代码" eorm:"跨越地物代码"`
	Q跨越地物类型        string `json:"跨越地物类型" eorm:"跨越地物类型"`
	Q跨越地物名称        string `json:"跨越地物名称" eorm:"跨越地物名称"`
	Q通航等级          string `json:"通航等级" eorm:"通航等级"`
	Q墩台防撞设施类型      string `json:"墩台防撞设施类型" eorm:"墩台防撞设施类型"`
	Q是否互通立交        string `json:"是否互通立交" eorm:"是否互通立交"`
	Q建设单位          string `json:"建设单位" eorm:"建设单位"`
	Q设计单位          string `json:"设计单位" eorm:"设计单位"`
	Q施工单位          string `json:"施工单位" eorm:"施工单位"`
	Q监理单位          string `json:"监理单位" eorm:"监理单位"`
	Q修建年度          string `json:"修建年度" eorm:"修建年度"`
	Q建成通车日期        string `json:"建成通车日期" eorm:"建成通车日期"`
	Q管养单位性质代码      string `json:"管养单位性质代码" eorm:"管养单位性质代码"`
	Q管养单位名称        string `json:"管养单位名称" eorm:"管养单位名称"`
	Q监管单位名称        string `json:"监管单位名称" eorm:"监管单位名称"`
	Q收费性质代码        string `json:"收费性质代码" eorm:"收费性质代码"`
	Q收费性质          string `json:"收费性质" eorm:"收费性质"`
	Q评定等级代码        string `json:"评定等级代码" eorm:"评定等级代码"`
	Q评定等级          string `json:"评定等级" eorm:"评定等级"`
	Q评定日期          string `json:"评定日期" eorm:"评定日期"`
	Q评定单位          string `json:"评定单位" eorm:"评定单位"`
	Q最后一次改造年度      string `json:"最后一次改造年度" eorm:"最后一次改造年度"`
	Q最后一次改造完工日期    string `json:"最后一次改造完工日期" eorm:"最后一次改造完工日期"`
	Q最后一次改造部位      string `json:"最后一次改造部位" eorm:"最后一次改造部位"`
	Q最后一次改造工程性质    string `json:"最后一次改造工程性质" eorm:"最后一次改造工程性质"`
	Q最后一次改造施工单位    string `json:"最后一次改造施工单位" eorm:"最后一次改造施工单位"`
	Q最后一次改造是否部补助项目 string `json:"最后一次改造是否部补助项目" eorm:"最后一次改造是否部补助项目"`
	Q当前主要病害部位      string `json:"当前主要病害部位" eorm:"当前主要病害部位"`
	Q当前主要病害        string `json:"当前主要病害" eorm:"当前主要病害"`
	Q交通管制措施代码      string `json:"交通管制措施代码" eorm:"交通管制措施代码"`
	Q交通管制措施        string `json:"交通管制措施" eorm:"交通管制措施"`
	Q所在政区代码        string `json:"所在政区代码" eorm:"所在政区代码"`
	Q桥梁所在位置        string `json:"桥梁所在位置" eorm:"桥梁所在位置"`
	Q是否宽路窄桥        string `json:"是否宽路窄桥" eorm:"是否宽路窄桥"`
	Q是否在长大桥梁目录中    string `json:"是否在长大桥梁目录中" eorm:"是否在长大桥梁目录中"`
	Q备注            string `json:"备注" eorm:"备注"`
}